API review: move avocado.settings to avocado.core.settings
The avocado.settings module is not intended to be used by tests but
by the framework itself. So, let's move it to avocado.core.
