+++ avocado/core/data_dir.py
@@ -34,9 +34,9 @@ import time
 import tempfile
 
 from avocado.core import job_id
+from avocado.core.settings import settings
 from avocado.utils import path as utils_path
 from avocado.utils.data_structures import Borg
-from avocado.settings import settings
 
 
 _BASE_DIR = os.path.join(sys.modules[__name__].__file__, "..", "..", "..")
