+++ avocado/core/restclient/cli/app.py
@@ -22,7 +22,7 @@ import types
 import importlib
 import functools
 
-from avocado import settings
+from avocado.core import settings
 from avocado.core import output
 from avocado.core import exit_codes
 from avocado.core.restclient import connection
